+++ src/simmate/website/workflows/migrations/0001_initial.py
@@ -8,7 +8,7 @@ class Migration(migrations.Migration):
     initial = True
 
     dependencies = [
-        ("core_components", "0002_alter_fingerprint_id_alter_fingerprintpool_id"),
+        ("core", "0002_alter_fingerprint_id_alter_fingerprintpool_id"),
     ]
 
     operations = [
@@ -144,7 +144,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],
@@ -218,7 +218,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],
@@ -505,7 +505,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],
@@ -726,7 +726,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
                 (
@@ -795,7 +795,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],
@@ -816,7 +816,7 @@ class Migration(migrations.Migration):
                 blank=True,
                 null=True,
                 on_delete=django.db.models.deletion.PROTECT,
-                to="core_components.spacegroup",
+                to="core.spacegroup",
             ),
         ),
         migrations.CreateModel(
@@ -889,7 +889,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],
@@ -964,7 +964,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],
@@ -1043,7 +1043,7 @@ class Migration(migrations.Migration):
                         blank=True,
                         null=True,
                         on_delete=django.db.models.deletion.PROTECT,
-                        to="core_components.spacegroup",
+                        to="core.spacegroup",
                     ),
                 ),
             ],