+++ docs/change_log.md
@@ -120,7 +120,8 @@ There is one significant exception to the above rules -- `MAJOR=0` releases. Any
 - update `breadcrumbs` tag to include deprec `breadcrumb_active`
 - workflows results in the web ui now exists within the Data tab. The workflow tab has been limited to submission, docs, and analytics
 - `DynamicApiView` was split up into separate views for easier debugging and future enhancements
-- `website.engine` apps and others are condensed into `core_components`, `data_explorer`, and `workflows` apps
+- renamed `website.core` to `website.server` and `website.core_components` to `website.core` to better align with standard Django architectural conventions
+- `website.engine` apps and others are condensed into `core`, `data_explorer`, and `workflows` apps
 - refactored `toolkit.featurizers` module
 - deprec django-unicorn in favor of htmx
 - deprec badelf app as it is now a standalone python package (baderkit)